Rename apollo-cache-control-js  and apollo-tracing-js and unify configs
