--- packages/apollo-cache-control-js/.npmignore
@@ -1,1 +1,0 @@
-__test__
