--- packages/apollo-cache-control-js/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "declaration": true,
-    "sourceMap": true,
-    "outDir": "dist",
-    "lib": ["es2017", "esnext.asynciterable"],
-    "types": ["jest", "node"],
-    "strict": true,
-    "noUnusedParameters": true,
-    "noUnusedLocals": true
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", "__tests__"]
-}
