+++ packages/apollo-cache-control/.npmignore
@@ -1,0 +1,6 @@
+*
+!src/**/*
+!dist/**/*
+dist/**/*.test.*
+!package.json
+!README.md
