+++ packages/apollo-server-memcached/package.json
@@ -47,7 +47,7 @@
       "js",
       "json"
     ],
-    "testRegex": "/__tests__/.*$",
+    "testRegex": "src/__tests__/.*$",
     "globals": {
       "ts-jest": {
         "skipBabel": true
