+++ packages/apollo-server-memcached/src/__tests__/Memcached.test.ts
@@ -1,7 +1,7 @@
 // use mock implementations for underlying databases
 jest.mock('memcached', () => require('memcached-mock'));
 
-import { MemcachedCache } from '../src/index';
-import { testKeyValueCache } from '../../apollo-server-caching/src/__tests__/testsuite';
+import { MemcachedCache } from '../index';
+import { testKeyValueCache } from '../../../apollo-server-caching/src/__tests__/testsuite';
 
 testKeyValueCache(new MemcachedCache('localhost'));
