+++ packages/apollo-server-redis/src/__tests__/Redis.test.ts
@@ -2,7 +2,7 @@
 jest.mock('redis', () => require('redis-mock'));
 jest.useFakeTimers(); // mocks out setTimeout that is used in redis-mock
 
-import { RedisCache } from '../src/index';
-import { testKeyValueCache } from '../../apollo-server-caching/src/__tests__/testsuite';
+import { RedisCache } from '../index';
+import { testKeyValueCache } from '../../../apollo-server-caching/src/__tests__/testsuite';
 
 testKeyValueCache(new RedisCache({ host: 'localhost' }));
