--- packages/apollo-tracing-js/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "declaration": true,
-    "sourceMap": true,
-    "outDir": "dist",
-    "lib": ["es2017", "esnext.asynciterable"],
-    "types": ["jest", "node"],
-    "strict": true
-  },
-  "include": ["src/**/*"]
-}
