+++ packages/apollo-tracing/package.json
@@ -1,45 +1,40 @@
 {
-  "name": "apollo-cache-control",
-  "version": "0.1.1",
-  "description": "A GraphQL extension for cache control",
+  "name": "apollo-tracing",
+  "version": "0.2.0-beta.1",
+  "description": "Collect and expose trace data for GraphQL requests",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "scripts": {
     "clean": "rm -rf dist",
     "compile": "tsc",
-    "prepublish": "npm run clean && npm run compile",
-    "test": "jest",
+    "prepublish": "npm run compile",
+    "test": "jest --verbose",
     "watch": "tsc -w"
   },
   "license": "MIT",
-  "repository": "apollographql/apollo-cache-control-js",
+  "repository": "apollographql/apollo-tracing-js",
   "author": "Martijn Walraven <martijn@martijnwalraven.com>",
   "engines": {
-    "node": ">=6.0"
+    "node": ">=4.0"
   },
   "dependencies": {
+    "apollo-server-env": "^2.0.0-rc.6",
     "graphql-extensions": "0.1.0-beta.0"
   },
   "peerDependencies": {
     "graphql": "0.10.x - 0.13.x"
   },
   "devDependencies": {
-    "@types/graphql": "^0.13.1",
     "@types/jest": "^23.1.2",
-    "@types/node": "^10.3.6",
-    "graphql": "^0.13.2",
-    "graphql-tools": "^3.0.4",
     "jest": "^23.2.0",
     "jest-matcher-utils": "^23.2.0",
-    "ts-jest": "^22.4.6",
-    "typescript": "^2.9.2"
-  },
-  "prettier": {
-    "trailingComma": "all",
-    "singleQuote": true
+    "ts-jest": "^22.4.6"
   },
   "jest": {
     "testEnvironment": "node",
+    "setupFiles": [
+      "<rootDir>/node_modules/apollo-server-env/dist/index.js"
+    ],
     "transform": {
       "^.+\\.(ts|js)$": "ts-jest"
     },
@@ -48,15 +43,7 @@
       "js",
       "json"
     ],
-    "testRegex": "apollo-cache-control-js/src/__tests__/.*$",
-    "testPathIgnorePatterns": [
-      "<rootDir>/node_modules/",
-      "<rootDir>/lib/",
-      "test-utils"
-    ],
-    "roots": [
-      "../../"
-    ],
+    "testRegex": "src/__tests__/.*$",
     "globals": {
       "ts-jest": {
         "skipBabel": true
